django/requirements.txt
django
django-click
django-extensions
django-types

ipython
django-admin-extra-buttons
djangorestframework
